+++ src/main/java/com/foolsix/fancyenchantments/enchantment/LavaBurst.java
@@ -4,12 +4,17 @@ import com.foolsix.fancyenchantments.FancyEnchantments;
 import com.foolsix.fancyenchantments.enchantment.EssentiaEnch.FEBaseEnchantment;
 import com.foolsix.fancyenchantments.enchantment.util.EnchUtils;
 import com.foolsix.fancyenchantments.util.ModConfig;
+import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.Style;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
-import net.minecraftforge.event.entity.living.LivingHurtEvent;
+import net.minecraft.world.level.Explosion;
+import net.minecraft.world.level.Level;
 
 public class LavaBurst extends FEBaseEnchantment {
     public static final String NAME = "lava_burst";
@@ -28,7 +33,14 @@ public class LavaBurst extends FEBaseEnchantment {
         return name;
     }
 
-    public void hitExplosion(LivingHurtEvent e) {
-        //todo
+    @Override
+    public void doPostAttack(LivingEntity pAttacker, Entity pTarget, int pLevel) {
+        if (Math.random() < CONFIG.probability * pLevel) {
+            Level level = pTarget.getLevel();
+            if (level instanceof ServerLevel) {
+                level.explode(pAttacker, pTarget.getX(), pTarget.getY(), pTarget.getZ(), 0.5f * pLevel, Explosion.BlockInteraction.NONE);
+                EnchUtils.generateSimpleParticleAroundEntity(pTarget, ParticleTypes.LAVA);
+            }
+        }
     }
 }