+++ src/main/java/com/foolsix/fancyenchantments/enchantment/Melter.java
@@ -19,6 +19,7 @@ public class Melter extends IgnisEnchantment {
     public static final String NAME = "melter";
     private static final ModConfig.MelterOptions CONFIG = FancyEnchantments.getConfig().melterOptions;
     private static final java.util.UUID ID = UUID.fromString("252d00fc-2176-a4d6-94bc-55a539753d45");
+
     public Melter() {
         super(CONFIG, EnchantmentCategory.WEAPON, new EquipmentSlot[]{EquipmentSlot.MAINHAND});
     }